Download individual tarballs instead of aggregate
Previously, we would download a prebuilt aggregated tarball from
zenoss-pip, augment it, then build the image from the augemented
tarball.  After this change we skip the aggregated tarball, downloading
the source tarballs and building the augmented aggregated tarball from
those.

The steps for building the aggregated tarball were moved from
opentsdb/makefile to hbase/makefile.  References to the augmented
tarball are changed to refer to the aggregated tarball, since there is
no longer a need to distinguish augmented from aggregated (because there
is no longer an unaugmented aggregated tarball)