Minimize artifacts in build directory
Previously intermediate artifacts were located in the build
subdirectory; after this change only those artifacts necessary for
building the image are in build.  Intermediate files are in the stage
subdirectory.

When building the image, the build sudbirectory is used for context;
i.e., the Dockerfile is copied to build and docker runs on that.