Normalize libhadoop.so handling
Formerly we would download libhadoop.so from zenpip and manually install
it into the hbase image; however, since we added hadoop to the image and
libhadoop is included in the hadoop package, we no longer need to do so.

This is preferable in any case since the former libhadoop artifact was
not necessarily the version compatible with the most recent hadoop
library.