Properly handle intermediate images/container
Previously the container in which the HDFS is initialized was leaked by
the build process.  This change cleans up that container.

Additionally, the initial image (without the initilaized HDFS) was
created with the canonical HBase image name.  In the event of a failure
in the HDFS initialization an invalid image could be left behind.  This
change writes that initial image to a temporary image name, avoiding
that possibility.

+ Create initial image with temporary name.
+ Use variables to clarify use of temporary image and container.
+ Clean up temporary image and container.